back/requirements.txt
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.13
django-cors-headers==3.4.0
djangorestframework==3.11.0
env-file==2020.5.9
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.2
public==2019.4.13
pytz==2020.1
requests==2.24.0
smmap==3.0.4
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
values==2019.4.13
